requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.35
aiomysql==0.2.0
redis==5.0.1
aiohttp==3.9.1
beautifulsoup4==4.12.3
feedparser==6.0.11
celery==5.3.6
pydantic-settings>=2.5.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
aiosqlite==0.19.0

# ====================== LLM Agent 依赖 ======================
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
